Format update
Updated formatting.  MD is picky about blank lines.  Where we have bullet lists (and we have a lot of them) with a mix of bullets and code, it is necessary to have no blank lines between the bullets and code.  Minus a blank line the code indents correctly; add a blank line and the code shifts left.  Code is aligned now.